move proto to separate crate
