+++ examples/rust/Cargo.toml
@@ -11,7 +11,7 @@ clap = { version = "3.2.22", features = ["cargo", "derive"] }
 env_logger = "0.10.0"
 futures = "0.3.24"
 log = { version = "0.4.14", features = ["std"] }
-solana-geyser-grpc = { path = "../../solana-geyser-grpc" }
+solana-geyser-grpc = { path = "../../triton-grpc-geyser" }
 solana-geyser-grpc-client = { path = "../../rust" }
 tokio = { version = "1.21.2", features = ["rt-multi-thread", "macros", "time"] }
