+++ rust/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 bytes = "1.3.0"
 futures = "0.3.24"
 http = "0.2.8"
-solana-geyser-grpc = { path = "../solana-geyser-grpc" }
+solana-geyser-grpc = { path = "../triton-grpc-geyser" }
 thiserror = "1.0"
 tonic = { version = "0.8.2", features = ["gzip", "tls", "tls-roots"] }
