+++ triton-grpc-geyser/config-test.json
@@ -1,0 +1,13 @@
+{
+    "libpath": "../target/debug/libsolana_geyser_grpc.so",
+    "log": {
+        "level": "info"
+    },
+    "grpc": {
+        "address": "0.0.0.0:10000",
+        "channel_capacity": "100_000"
+    },
+    "prometheus": {
+        "address": "0.0.0.0:8999"
+    }
+}
