+++ triton-grpc-proto/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "triton-grpc-proto"
+version = "1.0.0+solana.1.15.2"
+authors = ["Triton One"]
+edition = "2021"
+
+[dependencies]
+prost = "0.11.0"
+tonic = "0.8.0"
+
+[build-dependencies]
+anyhow = "1.0.62"
+protobuf-src = "1.1.0"
+tonic-build = "0.8.2"
