+++ triton-grpc-proto/build.rs
@@ -1,0 +1,5 @@
+fn main() -> anyhow::Result<()> {
+    std::env::set_var("PROTOC", protobuf_src::protoc());
+    tonic_build::compile_protos("../proto/geyser.proto")?;
+    Ok(())
+}
